requirements.txt
opencv-python==4.5.5.62
Pillow==9.0.0